# projetoExtensaoWebII-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API projeto extensão Web II",
        "description": "Documentação da API do projeto de extensão Web II",
        "version": "1.0"
    },
    "paths": {
...
        "/api/donations": {
            "get": {
                "tags": [
                    "Doações"
                ],
                "summary": "Listar todas as doações",
                "operationId": "efd463fc838a5d0b3edcc89ecc8374ac",
                "responses": {
                    "200": {
                        "description": "Lista de doações"
                    }
                }
            },
            "post": {
                "tags": [
                    "Doações"
                ],
                "summary": "Criar uma nova doação",
                "operationId": "a83d56537f533c01a1daf0abefdec913",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "organization",
                                    "goal_amount"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ajuda para Enchentes"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Campanha para ajudar vítimas de enchentes"
                                    },
                                    "organization": {
                                        "type": "string",
                                        "example": "Cruz Vermelha"
                                    },
                                    "goal_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Doação criada com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/donations/{id}": {
            "get": {
                "tags": [
                    "Doações"
                ],
                "summary": "Obter uma doação específica",
                "operationId": "d70f7a1ae033692540b40f026aa5740e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da doação"
                    },
                    "404": {
                        "description": "Doação não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Doações"
                ],
                "summary": "Atualizar uma doação",
                "operationId": "34b621cd5de718d266cdb93e0c5a0294",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ajuda para Enchentes"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Campanha para ajudar vítimas de enchentes"
                                    },
                                    "organization": {
                                        "type": "string",
                                        "example": "Cruz Vermelha"
                                    },
                                    "goal_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doação atualizada com sucesso"
                    },
                    "404": {
                        "description": "Doação não encontrada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Doações"
                ],
                "summary": "Excluir uma doação",
                "operationId": "ff34df7029ed4a4fc68e6a59674f9621",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doação excluída com sucesso"
                    },
                    "404": {
                        "description": "Doação não encontrada"
                    }
                }
            }
        },
        "/api/donations/{id}/contribute": {
            "post": {
                "tags": [
                    "Doações"
                ],
                "summary": "Contribuir para uma doação",
                "operationId": "917f85642d481da99b2c689677f02de0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contribuição realizada com sucesso"
                    },
                    "404": {
                        "description": "Doação não encontrada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
...
    },
...
}